% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgsPowerCalc_helper.R
\name{getCaseConBurdenPower}
\alias{getCaseConBurdenPower}
\title{Calculate statistical power in a case control burden analysis.}
\usage{
getCaseConBurdenPower(R, N, r, q, f, p_thres_cc_burden)
}
\arguments{
\item{R}{Relative Risk (numeric).}

\item{N}{Sample Size (integer).}

\item{r}{Case:Control ratio (numeric).}

\item{q}{Total number of variants within selected regions per control individual (numeric).}

\item{f}{Percent of variants within selected regions that mediate disease risk (numeric).}

\item{p_thres_cc_burden}{p-value threshold after correction for multiple comparisons (numeric, 0-1).}
}
\value{
The statistical power in a case control cohort.
}
\description{
Calculate statistical power in a case control burden analysis.
}
\examples{
getCaseConBurdenPower(2.5, 20000, 1, 1.26, 0.2, 5e-5)
}
