% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgsPowerCalc_helper.R
\name{getDeNovoBurdenPower}
\alias{getDeNovoBurdenPower}
\title{Calculate statistical power in a de novo burden analysis.}
\usage{
getDeNovoBurdenPower(R, N, q, f)
}
\arguments{
\item{R}{Relative Risk (numeric).}

\item{N}{Sample Size (integer).}

\item{q}{Total number of variants within selected regions per control individual (numeric).}

\item{f}{Percent of variants within selected regions that mediate disease risk (numeric).}
}
\value{
The statistical power in a de novo cohort.
}
\description{
Calculate statistical power in a de novo burden analysis.
}
\examples{
getDeNovoBurdenPower(25, 5000, 0.0249, 0.2)
}
