% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgsPowerCalc_helper.R
\name{getMultiVarLocusPower}
\alias{getMultiVarLocusPower}
\title{Calculates the power of rare variant locus discovery across a gene in cases and controls.}
\usage{
getMultiVarLocusPower(f, AF, sum_var, R, K, N, r, p_thres_cc_locus_multi)
}
\arguments{
\item{f}{Proportion of variants that mediate risk (numeric, 0-1).}

\item{AF}{Allele frequency (numeric, 0-1).}

\item{sum_var}{The sum of all heterozygous allele freuqencies (p * 1-p) for each replicates (numeric)}

\item{R}{Relative Risk (numeric).}

\item{K}{Prevalence of the disorder (numeric, 0-1).}

\item{N}{Sample Size (integer).}

\item{r}{Case:Control ratio (numeric).}

\item{p_thres_cc_locus_multi}{p-value threshold after correction for multiple comparisons (numeric, 0-1).}
}
\value{
The statistical power (numeric, 0-1).
}
\description{
Calculates the power of rare variant locus discovery across a gene in cases and controls.
}
\examples{
getMultiVarLocusPower(0.2, 1e-4, 0.6, 2.5, 0.01, 20000, 1, 2.5e-06)
}
