% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgsPowerCalc_helper.R
\name{getSingleVarLocusPower}
\alias{getSingleVarLocusPower}
\title{Calculates the power of rare variant locus discovery in cases and controls.}
\usage{
getSingleVarLocusPower(f, AF, R, K, N, r, p_thres_cc_locus_single)
}
\arguments{
\item{f}{Proportion of variants that mediate risk (numeric, 0-1).}

\item{AF}{Allele frequency (numeric, 0-1).}

\item{R}{Relative Risk (numeric).}

\item{K}{Prevalence of the disorder (numeric, 0-1).}

\item{N}{Sample Size (integer).}

\item{r}{Case:Control ratio (numeric).}

\item{p_thres_cc_locus_single}{p-value threshold after correction for multiple comparisons (numeric, 0-1).}
}
\value{
The statistical power (numeric, 0-1).
}
\description{
Calculates the power of rare variant locus discovery in cases and controls.
}
\examples{
getSingleVarLocusPower(0.2, 1e-4, 2.5, 0.01, 20000, 1, 1.666667e-11)
}
