% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgsPowerCalc_helper.R
\name{NCPgen}
\alias{NCPgen}
\title{Calculates the chi-square distribution with the non-centrality parameter χ2 in cases and controls.}
\usage{
NCPgen(q, R, K, CaseN, ConN)
}
\arguments{
\item{q}{Allele frequency (numeric, 0-1).}

\item{R}{Relative Risk (numeric).}

\item{K}{Prevalence of the disorder (numeric, 0-1).}

\item{CaseN}{Number of cases (integer).}

\item{ConN}{Number of controls (integer).}
}
\value{
The chi-squared value (numeric).
}
\description{
Calculates the chi-square distribution with the non-centrality parameter χ2 in cases and controls.
}
\examples{
NCPgen(1e-05, 2.5, 0.01, 10000, 10000)
}
