% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgsPowerCalc_helper.R
\name{plotCcLocusSingleByRelativeRisk}
\alias{plotCcLocusSingleByRelativeRisk}
\title{Tests the power to detect a single locus using case control data varying relative risk.}
\usage{
plotCcLocusSingleByRelativeRisk(R, N, r = 1, s, f_gene, f, K, AF_bar, N_rep,
  p_thres_cc_locus_single = 0.05, name = "Test", col = "#e41a1c")
}
\arguments{
\item{R}{Maximum relative risk (numeric).}

\item{N}{Sample size (integer).}

\item{r}{Case:Control ratio (numeric).}

\item{s}{Number of possible rare variants per gene (numeric).}

\item{f_gene}{Number of possible functional rare variants per gene (numeric).}

\item{f}{Proportion of functional rare variants that mediate risk for the disorder (numeric, 0-1).}

\item{K}{Prevalence of the disorder (numeric, 0-1).}

\item{AF_bar}{Average minor allele frequency (numeric, 0-1).}

\item{N_rep}{Number of replicate simulations (integer).}

\item{p_thres_cc_locus_single}{p-value threshold after correction for multiple comparisons (numeric, 0-1).}

\item{name}{Prefix for plot filenames (text).}

\item{col}{Color of the line (hex).}
}
\value{
A PDF showing the estimated statistical power as the relative risk is varied.
}
\description{
Tests the power to detect a single locus using case control data varying relative risk.
}
\examples{
R <- 2.5
N <- 20000
r <- 1
s <- 615
f_gene <- 123
f <- 0.2
K <- 0.01
AF_bar <- 0.001
N_rep <- 50
p_thres_cc_locus_single <- 0.05 / 3000000000
name <- "Test1"
col <- "#e41a1c"
plotCcLocusSingleByRelativeRisk()
}
