% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgsPowerCalc_helper.R
\name{plotCcLocusSingleBySampleSize}
\alias{plotCcLocusSingleBySampleSize}
\title{Tests the power to detect a single locus using case control data varying sample size}
\usage{
plotCcLocusSingleBySampleSize(f_gene, s, AF_bar, N_rep, R, f, K, N, r,
  p_thres_cc_locus_single = 0.05, name = "Test", col = "#e41a1c")
}
\arguments{
\item{f_gene}{Number of possible functional rare variants per gene (numeric).}

\item{s}{Number of possible rare variants per gene (numeric).}

\item{AF_bar}{Average minor allele frequency (numeric, 0-1).}

\item{N_rep}{Number of replicate simulations (integer).}

\item{R}{Relative risk (numeric).}

\item{f}{Proportion of functional rare variants that mediate risk for the disorder (numeric, 0-1).}

\item{K}{Prevalence of the disorder (numeric, 0-1).}

\item{N}{Maximum sample size (integer).}

\item{r}{Case:Control ratio (numeric).}

\item{p_thres_cc_locus_single}{p-value threshold after correction for multiple comparisons (numeric, 0-1).}

\item{name}{Prefix for plot filenames (text).}

\item{col}{Color of the line (hex).}
}
\value{
A PDF showing the estimated statistical power as the sample size is varied.
}
\description{
Tests the power to detect a single locus using case control data varying sample size
}
\examples{
f_gene <- 123
s <- 615
AF_bar <- 0.001
N_rep <- 50
R <- 2.5
f <- 0.2
K <- 0.01
N <- 20000
r <- 1
p_thres_cc_locus_single <- 0.05 / 3000000000
name <- "Test1"
col <- "#e41a1c"
plotCcLocusSingleBySampleSize()
}
