% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgsPowerCalc_helper.R
\name{plotDnBurdenByRelativeRisk}
\alias{plotDnBurdenByRelativeRisk}
\title{Tests the power to detect a burden difference using de novo data varying relative risk.}
\usage{
plotDnBurdenByRelativeRisk(R, N, q, f, p_thres_denovo_burden = 0.05,
  name = "Test", col = "#e41a1c")
}
\arguments{
\item{R}{Maximum relative risk (numeric).}

\item{N}{Sample size (integer).}

\item{q}{Number of functional rare variants per person in seleceted regions (integer).}

\item{f}{Proportion of functional rare variants that mediate risk for the disorder (numeric, 0-1).}

\item{p_thres_denovo_burden}{p-value threshold after correction for multiple comparisons (numeric, 0-1).}

\item{name}{Prefix for plot filenames (text).}

\item{col}{Color of the line (hex).}
}
\value{
A PDF showing the estimated statistical power as the relative risk is varied.
}
\description{
Tests the power to detect a burden difference using de novo data varying relative risk.
}
\examples{
R <- 25
N <- 5000
q <- 0.0249
f <- 0.2
p_thres_denovo_burden <- 0.05 / 1000
name <- "Test1"
col <- "#e41a1c"
plotDnBurdenByRelativeRisk()
}
