% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgsPowerCalc_helper.R
\name{getDeNovoBurdenPower}
\alias{getDeNovoBurdenPower}
\title{Calculate statistical power in a de novo burden analysis.}
\usage{
getDeNovoBurdenPower(R, N, r, q, f, p_thres_denovo_burden)
}
\arguments{
\item{R}{Relative Risk (numeric).}

\item{N}{Sample Size (integer).}

\item{r}{Case:Control ratio (numeric).}

\item{q}{Total number of variants within selected regions per control individual (numeric).}

\item{f}{Percent of variants within selected regions that mediate disease risk (numeric).}

\item{p_thres_denovo_burden}{p-value threshold after correction for multiple comparisons (numeric, 0-1).}
}
\value{
The statistical power in a de novo cohort.
}
\description{
Calculate statistical power in a de novo burden analysis.
}
\examples{
getDeNovoBurdenPower(25, 5000, 1, 0.0249, 0.2, 5e-5)
}
